Keep the umbrella sdist's tests to files it ships
The umbrella sdist's only-include is src/nab, tests and conftest.py, so docs/ is
not in it, but three shipped test modules check the CLI against
docs/reference/cli.md and docs/explanation/conflicts.md. Building the sdist and
running `pytest tests` from the extraction fails 16 tests on FileNotFoundError.

Those 16 move to tests/test_cli_docs.py, which the sdist leaves out the way it
already leaves out the tests that read docs/contributing.md, tasks/, the noxfile
and the workflows. Excluding test_cli.py wholesale would have cost the sdist the
309 CLI tests that sit alongside the 4 in it that read a page. hermetic_roots
moves to the suite's conftest, which both modules now use.

tests/test_sdist_manifest.py holds the line: the modules the suite has that
build a path outside the sdist have to be exactly the ones it leaves out. It
reads string path components, so a module placed outside only by a computed
component would be invisible to it; the second check catches that case and
ignores a computed component on a module already placed by a path it can read.
