+++ tests/test_cli.py
@@ -3621,249 +3621,6 @@ class TestConfigErrors:
         assert "workspace discovery error" in capsys.readouterr().err
 
 
-_CONFLICTS_DOC = (
-    Path(__file__).resolve().parents[1] / "docs" / "explanation" / "conflicts.md"
-)
-
-
-def _console_block(text: str, command: str) -> list[str]:
-    """The output lines of the ``console`` block whose first line is ``command``."""
-    for body in re.findall(r"```console\n(.*?)\n```", text, re.DOTALL):
-        lines = body.splitlines()
-        if lines[0] == command:
-            return lines[1:]
-
-    msg = f"no console block for {command!r}"
-    raise AssertionError(msg)
-
-
-def _doc_paragraph(text: str, needle: str) -> str:
-    """The blank-line-delimited paragraph of ``text`` that contains ``needle``."""
-    for paragraph in text.split("\n\n"):
-        if needle in paragraph:
-            return paragraph
-
-    msg = f"no paragraph containing {needle!r}"
-    raise AssertionError(msg)
-
-
-def _emitted_labels(
-    tmp_path: Path,
-    capsys: pytest.CaptureFixture[str],
-    body: str,
-    *,
-    extras: tuple[str, ...] = (),
-    groups: tuple[str, ...] = (),
-) -> list[str]:
-    """The ``# label`` headers a ``nab lock`` over ``body`` prints for the selection.
-
-    The requirements formats label one block per emitted target and omit
-    the header when there is only one, so for these single-target
-    specific-mode locks the headers count the forks.
-    """
-    pyproject = _make_pyproject(tmp_path, body)
-    lock(
-        pyproject,
-        cache_dir=tmp_path / "cache",
-        offline=True,
-        extras=extras,
-        groups=groups,
-        format="requirements-without-hashes",
-        output=Path("-"),
-    )
-
-    printed = capsys.readouterr().out
-    return [line for line in printed.splitlines() if line.startswith("# ")]
-
-
-class TestConflictsDocTranscript:
-    """The conflicts page quotes the refusal lines the CLI prints."""
-
-    _UMBRELLA = (
-        '[project]\nname = "proj"\nversion = "0.1.0"\ndependencies = []\n'
-        "[project.optional-dependencies]\n"
-        "cpu = []\n"
-        "gpu = []\n"
-        'all = ["proj[cpu]", "proj[gpu]"]\n'
-        "[tool.nab]\n"
-        'conflicts = [[{ extra = "cpu" }, { extra = "gpu" }]]\n'
-    )
-
-    _DEFAULT_GROUPS = (
-        '[project]\nname = "proj"\nversion = "0.1.0"\ndependencies = []\n'
-        "[dependency-groups]\n"
-        'black22 = ["black==22.1.0"]\n'
-        'black23 = ["black==23.12.0"]\n'
-        "[tool.nab]\n"
-        'default-groups = ["black22", "black23"]\n'
-        'conflicts = [[{ group = "black22" }, { group = "black23" }]]\n'
-    )
-
-    def test_umbrella_refusal_matches_documented_transcript(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """An umbrella extra reaching both members prints the page's line."""
-        pyproject = _make_pyproject(tmp_path, self._UMBRELLA)
-        with pytest.raises(SystemExit, match="1"):
-            lock(pyproject, cache_dir=tmp_path / "cache", extras=("all",))
-        printed = capsys.readouterr().err.strip()
-
-        doc = _CONFLICTS_DOC.read_text(encoding="utf-8")
-        assert printed in _console_block(doc, "$ nab lock --extras all")
-
-    def test_default_groups_refusal_matches_documented_transcript(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """Two default groups in one exclusive set print the page's line."""
-        pyproject = _make_pyproject(tmp_path, self._DEFAULT_GROUPS)
-        with pytest.raises(SystemExit, match="1"):
-            lock(pyproject, cache_dir=tmp_path / "cache")
-        printed = capsys.readouterr().err.strip()
-
-        doc = _CONFLICTS_DOC.read_text(encoding="utf-8")
-        assert printed in _console_block(doc, "$ nab lock")
-
-
-class TestConflictsDocForkingPolicies:
-    """The conflicts page names which policies fork a co-selecting run.
-
-    Only the exclusive policies do; ``at-least-one`` permits co-selection.
-    """
-
-    _GROUPS = (
-        '[project]\nname = "proj"\nversion = "0.1.0"\ndependencies = []\n'
-        "[dependency-groups]\n"
-        "a = []\n"
-        "b = []\n"
-    )
-
-    def _body(self, policy: str) -> str:
-        """The two-group project with ``a`` and ``b`` conflicting under ``policy``."""
-        members = '[{ group = "a" }, { group = "b" }]'
-        return (
-            self._GROUPS
-            + "[tool.nab]\n"
-            + f'conflicts = [{{ members = {members}, policy = "{policy}" }}]\n'
-        )
-
-    def _unwrapped_claim(self, needle: str) -> str:
-        """The page's paragraph holding ``needle``, rejoined onto one line.
-
-        The page hard-wraps, so a claim spanning a line break only matches unwrapped.
-        """
-        doc = _CONFLICTS_DOC.read_text(encoding="utf-8")
-        return " ".join(_doc_paragraph(doc, needle).split())
-
-    def test_exactly_one_forks_co_selected_members(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """An exclusive set resolves each co-selected member on its own."""
-        labels = _emitted_labels(
-            tmp_path, capsys, self._body("exactly-one"), groups=("a", "b")
-        )
-        assert labels == ["# host-group-a", "# host-group-b"]
-
-        claim = self._unwrapped_claim("When the selection activates")
-        assert "an exclusive set (`at-most-one` or `exactly-one`)" in claim
-
-    def test_at_least_one_co_selection_stays_one_resolve(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """An at-least-one set permits co-selection, so the run does not fork."""
-        labels = _emitted_labels(
-            tmp_path, capsys, self._body("at-least-one"), groups=("a", "b")
-        )
-        assert labels == []
-
-        claim = self._unwrapped_claim("The require-one policies")
-        assert "`at-least-one` permits it" in claim
-
-
-_CLI_REFERENCE_DOC = (
-    Path(__file__).resolve().parents[1] / "docs" / "reference" / "cli.md"
-)
-
-
-def _doc_section(text: str, heading: str) -> str:
-    return text.partition(f"\n{heading}\n")[2].partition("\n## ")[0]
-
-
-def _names_flag(text: str, flag: str) -> bool:
-    """Whether ``text`` names ``flag``, its ``--no-`` form, or a covering wildcard."""
-    forms = [flag, f"--no-{flag.removeprefix('--')}"]
-    if flag.startswith("--project-"):
-        forms.append("--project-*")
-    return any(re.search(rf"`{re.escape(form)}(?![\w-])", text) for form in forms)
-
-
-class TestCliReferenceSelectionShape:
-    """The reference's selection paragraph matches how many resolves a selection runs.
-
-    ``--extras`` and ``--groups`` union into one resolve until they select
-    two members of a declared conflict set, which forks the run.
-    """
-
-    _EXTRAS = (
-        '[project]\nname = "proj"\nversion = "0.1.0"\ndependencies = []\n'
-        "[project.optional-dependencies]\n"
-        "cpu = []\n"
-        "gpu = []\n"
-    )
-
-    _CONFLICT = '[tool.nab]\nconflicts = [[{ extra = "cpu" }, { extra = "gpu" }]]\n'
-
-    def _selection_paragraph(self) -> str:
-        """The ``nab lock`` paragraph that states what a selection resolves to."""
-        text = _doc_section(
-            _CLI_REFERENCE_DOC.read_text(encoding="utf-8"), "## `nab lock`"
-        )
-        return _doc_paragraph(text, "union resolve")
-
-    def test_selection_alone_is_one_union_resolve(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """Two extras with no conflict declared resolve once, as the page says."""
-        labels = _emitted_labels(tmp_path, capsys, self._EXTRAS, extras=("cpu", "gpu"))
-        assert labels == []
-
-        assert "single union resolve" in self._selection_paragraph()
-
-    def test_co_selected_conflict_members_fork_the_resolve(
-        self, tmp_path: Path, capsys: pytest.CaptureFixture[str]
-    ) -> None:
-        """Declaring the same two extras exclusive resolves each separately."""
-        labels = _emitted_labels(
-            tmp_path, capsys, self._EXTRAS + self._CONFLICT, extras=("cpu", "gpu")
-        )
-        assert labels == ["# host-extra-cpu", "# host-extra-gpu"]
-
-        paragraph = self._selection_paragraph()
-        assert "`[tool.nab].conflicts`" in paragraph
-        assert "../explanation/conflicts.md" in paragraph
-
-
-class TestCliReferenceFlagCoverage:
-    """Each run subcommand's reference section names every flag it accepts."""
-
-    @pytest.mark.parametrize(
-        ("heading", "command"),
-        [("## `nab lock`", lock), ("## `nab download`", download)],
-    )
-    def test_section_names_every_flag(
-        self, heading: str, command: Callable[..., None]
-    ) -> None:
-        text = _CLI_REFERENCE_DOC.read_text(encoding="utf-8")
-
-        # Flags shared by both commands are documented once, in their own section.
-        scope = _doc_section(text, heading) + _doc_section(text, "## Runtime flags")
-
-        for name, param in inspect.signature(command).parameters.items():
-            if param.kind is not inspect.Parameter.KEYWORD_ONLY:
-                continue
-            flag = "--" + name.replace("_", "-")
-            assert _names_flag(scope, flag), f"{heading} omits {flag}"
-
-
 class TestCliDocstringCommandModules:
     """``nab.cli``'s docstring names every module that registers a subcommand."""
