+++ tests/test_output.py
@@ -5,7 +5,6 @@ from __future__ import annotations
 import io
 import logging
 import sys
-from pathlib import Path
 
 import pytest
 
@@ -476,77 +475,3 @@ def test_progress_repaints_after_diagnostic() -> None:
     printer.warning("careful")
     reporter.on_pin(1)
     assert err.getvalue().endswith("\r\033[K⠙ Resolving... 1 fetched, 1 pinned")
-
-
-_CLI_REFERENCE = Path(__file__).resolve().parents[1] / "docs" / "reference" / "cli.md"
-
-_SUBCOMMANDS = ("lock", "download", "config", "cache")
-
-
-def _cli_reference_text() -> str:
-    return _CLI_REFERENCE.read_text(encoding="utf-8")
-
-
-def _section_body(text: str, heading: str) -> str:
-    after = text.partition(f"\n{heading}\n")[2]
-    return after.partition("\n## ")[0]
-
-
-class TestCliReferenceDocumentsOutputPolicy:
-    """The CLI reference must list every output-policy flag and env var the CLI accepts.
-
-    ``parse_output_options`` defines the flags and ``Printer`` reads the env vars.
-    """
-
-    def test_verbosity_flags_documented(self) -> None:
-        text = _cli_reference_text()
-        for flag in ("-v", "-vv", "-q", "-qq", "--verbose", "--quiet"):
-            assert f"`{flag}`" in text, f"CLI reference omits verbosity flag {flag}"
-
-    def test_color_flags_documented(self) -> None:
-        text = _cli_reference_text()
-        assert "`--color`" in text
-        assert "`--no-color`" in text
-        for choice in ColorChoice:
-            assert f"`{choice.value}`" in text, (
-                f"CLI reference omits --color value {choice.value}"
-            )
-
-    def test_progress_documented(self) -> None:
-        text = _cli_reference_text()
-        assert "`--no-progress`" in text
-        assert "Resolving" in text
-
-    def test_output_env_vars_documented(self) -> None:
-        text = _cli_reference_text()
-        for var in ("NAB_VERBOSITY", "NAB_NO_PROGRESS", "NO_COLOR", "FORCE_COLOR"):
-            assert var in text, f"CLI reference omits env var {var}"
-
-    def test_nab_verbosity_values_documented(self) -> None:
-        text = _cli_reference_text()
-        for level in Verbosity:
-            name = level.name.lower()
-            assert f"`{name}`" in text, (
-                f"CLI reference omits NAB_VERBOSITY value {name!r}"
-            )
-
-    def test_output_control_scope_covers_every_subcommand(self) -> None:
-        """The scope paragraph must name every subcommand the flags reach.
-
-        ``main`` extracts a global ``-q`` before dispatching to any
-        subcommand, so the doc's enumeration must include ``cache``.
-        """
-        for sub in _SUBCOMMANDS:
-            _opts, rest = parse_output_options(["-q", sub], {})
-            assert rest == [sub], f"global -q not extracted before {sub!r}"
-
-        text = _cli_reference_text()
-        scope = next(
-            para
-            for para in _section_body(text, "## Output control").split("\n\n")
-            if "before the subcommand" in para
-        )
-        for sub in _SUBCOMMANDS:
-            assert f"`{sub}`" in scope, (
-                f"Output control scope omits the {sub!r} subcommand"
-            )
